src/workers/generate-and-store-hash/requirements.txt
boto3==1.14.51
botocore==1.17.51
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
docutils==0.15.2
dnspython==2.0.0
idna==2.10
imageio==2.9.0
imagehash==4.1.0
imageio-ffmpeg==0.4.2
imutils==0.5.3
jmespath==0.10.0
joblib==0.16.0
katna==0.4.1.4
kiwisolver==1.2.0
moviepy==1.0.3
mp3hash==0.1
networkx==2.5
numpy==1.19.1
opencv-contrib-python==4.4.0.42
pika==1.1.0
Pillow==7.2.0
proglog==0.1.9
pymongo==3.11.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.14.0
PyWavelets==1.1.1
requests==2.24.0
s3transfer==0.3.3
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.2
six==1.15.0
threadpoolctl==2.1.0
tifffile==2020.8.25
tqdm==4.48.2
urllib3==1.25.10
